Second to last update
* Well everyone, it was an amazing journey making this product.
However, due to us not wanting to release it, this app will not be
released.  We apologize for any inconvenience this may cause in
breaking any pinky promises we made at the beginning.
* Added a Mac App executable to the executables folder incase you
really want to run it on a Mac.  The database will be kept alive
indefinitely
* Added graph of full commit history.  It’s pretty sick
* Added our keynote.  I left it in keynote format to force you to get a
Mac or go to iCloud.com to see it.
* Added a PDF version because I really didn’t feel like hearing Alex
rant about how Apple will be stealing his information while he’s
viewing on the web.😉
* Some final tiny UI changes to Login, Alert, and Window to force
presentation on the second monitor.